Default to 12pt font in all environments.
Under OS X, this actually requires a bit of tweaking.  The issue is that
a point--when it comes to font sizes--is a bit more arbitrary under OS
X.  This was supposedly done to help with compatibility and resolution
scaling--though I disagree with this argument, so it ends up that the
way to think about point sizes under OS X is really as though it was on
a 72 dpi screen.  Since monitors are generally 96 dpi, we need to scale
up a bit.  While Retina displays have a much higher dpi, OS X is helping
to scale things, so it's still safe to assume a 96 dpi display.