Disable Syntastic's auto-location list from popping up.
When editing a file that has an "error" and trying to exit with `:wq`,
the location list will pop up and prevent vim from exiting.  This is
frustrating for two reasons: I'm not always in control of the file (it
may be some other project's code), and the settings may not match what's
being used by the project (I may be making a quick hack to try out an
idea).  Either way, I prefer the location list to not get in the way as
quick hacks are common for testing things.